+++ src/main/java/com/smday/fireworks/common/api/ResultCode.java
@@ -13,16 +13,16 @@ public enum ResultCode implements IErrorCode {
     FORBIDDEN(403, "没有相关权限");
     
 
-    private long code;
+    private int code;
     private String message;
 
-    private ResultCode(long code, String message) {
+    private ResultCode(int code, String message) {
         this.code = code;
         this.message = message;
     }
 
     @Override
-    public long getCode() {
+    public int getCode() {
         return code;
     }
